requirements.txt
Django
requests
urllib3
gunicorn
whitenoise
django-cors-headers
psycopg2-binary
python-dotenv